Reworked and consolidated data explorer closed UI
