+++ src/vs/workbench/browser/positronDataExplorer/positronDataExplorer.css
@@ -13,40 +13,3 @@
 	background-color: var(--vscode-positronDataExplorer-background);
 	grid-template-rows: [action-bar] 32px [data-explorer-panel] 1fr [end];
 }
-
-.positron-data-explorer-overlay {
-	top: 0;
-	left: 0;
-	right: 0;
-	bottom: 0;
-	z-index: 30;
-	display: grid;
-	position: absolute;
-	background-color: transparent;
-	grid-template-rows: [top-gutter] 1fr [message] max-content [bottom-gutter] 1fr [end];
-	grid-template-columns: [left-gutter] 1fr [message] max-content [right-gutter] 1fr [end];
-}
-
-.positron-data-explorer-overlay
-.message {
-	color: white;
-	display: flex;
-	row-gap: 10px;
-	font-size: 16px;
-	cursor: pointer;
-	padding: 25px 50px;
-	border-radius: 6px;
-	flex-direction: column;
-	grid-row: message / bottom-gutter;
-	grid-column: message / right-gutter;
-	box-shadow: 0 0 8px 2px var(--vscode-widget-shadow);
-	color: var(--vscode-positronModalDialog-foreground);
-	border: 1px solid var(--vscode-positronModalDialog-border);
-	background-color: var(--vscode-positronModalDialog-background);
-}
-
-.positron-data-explorer-overlay
-.message
-.message-line {
-	text-align: center;
-}
