+++ src/vs/workbench/browser/positronDataExplorer/positronDataExplorer.tsx
@@ -10,16 +10,15 @@ import * as React from 'react';
 import { PropsWithChildren, useEffect, useState } from 'react'; // eslint-disable-line no-duplicate-imports
 
 // Other dependencies.
-import { localize } from 'vs/nls';
 import { DisposableStore } from 'vs/base/common/lifecycle';
 import { ILayoutService } from 'vs/platform/layout/browser/layoutService';
 import { IClipboardService } from 'vs/platform/clipboard/common/clipboardService';
-import { PositronButton } from 'vs/base/browser/ui/positronComponents/button/positronButton';
 import { ActionBar } from 'vs/workbench/browser/positronDataExplorer/components/actionBar/actionBar';
 import { PositronActionBarServices } from 'vs/platform/positronActionBar/browser/positronActionBarState';
 import { PositronDataExplorerContextProvider } from 'vs/workbench/browser/positronDataExplorer/positronDataExplorerContext';
 import { DataExplorerPanel } from 'vs/workbench/browser/positronDataExplorer/components/dataExplorerPanel/dataExplorerPanel';
 import { IPositronDataExplorerInstance } from 'vs/workbench/services/positronDataExplorer/browser/interfaces/positronDataExplorerInstance';
+import { PositronDataExplorerClosed } from 'vs/workbench/browser/positronDataExplorer/components/dataExplorerClosed/positronDataExplorerClosed';
 
 /**
  * PositronDataExplorerServices interface.
@@ -82,35 +81,13 @@ export const PositronDataExplorer = (props: PropsWithChildren<PositronDataExplor
 			<div className='positron-data-explorer'>
 				<ActionBar />
 				<DataExplorerPanel />
-				{closed && (
-					<div className='positron-data-explorer-overlay'>
-						<PositronButton className='message' onPressed={props.onClose}>
-							<div
-								className='message-line'>
-								{(() => localize(
-									'positron.dataExplorer.dataDisplayName',
-									'{0} Data: {1}',
-									props.instance.languageName,
-									displayName
-								))()}
-							</div>
-							<div
-								className='message-line'>
-								{(() => localize(
-									'positron.dataExplorer.isNoLongerAvailable',
-									'Is no longer available'
-								))()}
-							</div>
-							<div
-								className='message-line'>
-								{(() => localize(
-									'positron.dataExplorer.clickToClose',
-									"Click to close Data Explorer"
-								))()}
-							</div>
-						</PositronButton>
-					</div>
-				)}
+				{closed &&
+					<PositronDataExplorerClosed
+						languageName={props.instance.languageName}
+						displayName={displayName}
+						onClose={props.onClose}
+					/>
+				}
 			</div>
 		</PositronDataExplorerContextProvider>
 	);
