+++ src/vs/workbench/contrib/positronDataExplorerEditor/browser/positronDataExplorerEditor.css
@@ -1,3 +1,7 @@
 /*---------------------------------------------------------------------------------------------
  *  Copyright (C) 2023-2024 Posit Software, PBC. All rights reserved.
  *--------------------------------------------------------------------------------------------*/
+
+.positron-data-explorer-container {
+	position: relative;
+}
